backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
anthropic==0.39.0
openai==1.54.0
pyyaml==6.0.1
tomli-w==1.0.0
packaging==23.2

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
